requirements.txt
# Core web framework
fastapi
#uvicorn
uvicorn[standard]
pandas

# LangChain and its components
openai
langchain
langchain-core
langchain-openai 
langchain-pinecone 
langchain-community

# Data loading and text processing
beautifulsoup4
pypdf # For loading PDF files

# Environment variable management
python-dotenv

# Database
pydantic-settings
sqlalchemy
psycopg2-binary
pgvector